requirements.txt
# LangChain 依赖包
# 安装命令: pip install -r requirements.txt

# LangChain 核心
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.10

# LLM 集成 (OpenAI 兼容接口)
langchain-openai>=0.0.2

# 向量存储与检索
chromadb>=0.4.0
faiss-cpu>=1.7.0

# 文档加载器
langchain-text-splitters>=0.0.1

# Embedding 模型
sentence-transformers>=2.2.0
huggingface-hub>=0.19.0
torch>=2.0.0

# PDF 加载 (可选)
pypdf>=3.0.0

# DOCX 加载
python-docx>=0.8.0

# 其他依赖
anthropic>=0.18.0
openai>=1.0.0

# 中文 NLP 处理
jieba>=0.42.1
